refactor(stark): unify owned/archived Table dimensions_consistent and into_frame
Give owned Table a dimensions_consistent method (reading length through
row_major_data() so it stays correct under disk-spill) and have both
StarkTableView arms delegate to it. Collapse the two verbatim into_frame
copies into a single StarkTableView::into_frame written over the uniform
get_row/height accessors, deleting the duplicated owned and archived
bodies. No behavior change.
